Feat: create and delete trainer routes
