+++ src/routes/index.ts
@@ -1,83 +1,8 @@
 import { Router } from 'express';
-import { prisma } from '../lib/prisma';
-import { z } from 'zod';
+import { pokemonRoutes } from './pokemon.routes';
+import { trainerRoutes } from './trainer.routes';
 
 export const routes = Router();
 
-routes.get('/pokemon', async (request, response) => {
-  const requestQuerySchema = z.object({
-    page: z.string().transform((p) => {
-      if (Number(p) <= 0) return 1;
-
-      return Number(p);
-    }),
-    generation: z.string().transform((gen) => {
-      const genNumber = Number(gen);
-      return isNaN(genNumber) || genNumber <= 0 ? undefined : genNumber;
-    }),
-  });
-
-  const { page, generation } = requestQuerySchema.parse(request.query);
-
-  const pokemons = await prisma.pokemon.findMany({
-    orderBy: {
-      pokedex_number: 'asc',
-    },
-    where: {
-      generation,
-    },
-    take: 20,
-    skip: (page - 1) * 20,
-  });
-
-  return response.json({
-    pokemons,
-  })
-});
-
-routes.post('/pokemon', async (request, response) => {
-  const requestBodySchema = z.object({
-    name: z.string(),
-    generation: z.number(),
-    pokedexNumber: z.number(),
-    type1: z.string(),
-    type2: z.string(),
-    legendary: z.number(),
-  });
-
-  const {
-    name,
-    generation,
-    pokedexNumber,
-    type1,
-    type2,
-    legendary,
-  } = requestBodySchema.parse(request.body);
-
-  const pokedexSlotTaken = await prisma.pokemon.findFirst({
-    where: {
-      pokedex_number: pokedexNumber,
-    }
-  });
-
-  if (pokedexSlotTaken) {
-    return response.status(400).json({
-      message: 'Pokedex slot already taken',
-    });
-  }
-
-  const pokemon = await prisma.pokemon.create({
-    data: {
-      name,
-      generation,
-      type_1: type1,
-      type_2: type2,
-      pokedex_number: pokedexNumber,
-      legendary,
-    }
-  });
-
-  return response.status(201).json({
-    pokemon,
-  })
-});
+routes.use('/pokemon', pokemonRoutes);
+routes.use('/trainer', trainerRoutes);
